requirements.txt
fastapi>=0.103.0,<0.104.0
uvicorn>=0.23.0,<0.24.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.0,<0.26.0
gunicorn>=21.2.0,<22.0.0
pytest>=7.4.0,<7.5.0
pytest-cov>=4.1.0,<4.2.0
black>=23.7.0,<23.8.0
flake8>=6.1.0,<6.2.0
isort>=5.12.0,<5.13.0
alembic>=1.12.0,<1.13.0
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.6,<2.9.7
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.5,<0.0.6
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
bcrypt==3.2.2
